oci: Force kill the container process only if nothing else worked
The function waitContainerStop() in oci/oci.go was designed to check
on a container process and wait for it until it returns, or returns
an error if a timeout was reached.

Now, looking at the way this function is used, it is called twice
from the same function StopContainer() with the idea of trying to
wait the process properly first. Then, if the process is not
terminating as expected, another signal is sent to the container
process, and another call to waitContainerStop() is issued.

This second call should be the only one calling into SIGKILL because
this is the last chance to stop the container process. But from the
first call into this function, we should only retrieve sort of a
boolean value telling if the process is still around, instead of
trying to kill it with a SIGKILL already.

This commit modifies waitContainerStop() by adding a simple flag
indicating if this function should try to force kill the container
process using SIGKILL signal in case the process is still around
after the timeout has been reached.

Fixes #1327

+++ oci/oci.go
@@ -561,7 +561,7 @@ func (r *Runtime) UpdateContainer(c *Container, res *rspec.LinuxResources) error
 	return nil
 }
 
-func waitContainerStop(ctx context.Context, c *Container, timeout time.Duration) error {
+func waitContainerStop(ctx context.Context, c *Container, timeout time.Duration, ignoreKill bool) error {
 	done := make(chan struct{})
 	// we could potentially re-use "done" channel to exit the loop on timeout,
 	// but we use another channel "chControl" so that we never panic
@@ -594,6 +594,10 @@ func waitContainerStop(ctx context.Context, c *Container, timeout time.Duration)
 		return ctx.Err()
 	case <-time.After(timeout):
 		close(chControl)
+		if ignoreKill {
+			return fmt.Errorf("failed to wait process, timeout reached after %.0f seconds",
+				timeout.Seconds())
+		}
 		err := unix.Kill(c.state.Pid, unix.SIGKILL)
 		if err != nil && err != unix.ESRCH {
 			return fmt.Errorf("failed to kill process: %v", err)
@@ -670,7 +674,7 @@ func (r *Runtime) StopContainer(ctx context.Context, c *Container, timeout int64
 		if err := utils.ExecCmdWithStdStreams(os.Stdin, os.Stdout, os.Stderr, r.Path(c), "kill", c.id, c.GetStopSignal()); err != nil {
 			return fmt.Errorf("failed to stop container %s, %v", c.id, err)
 		}
-		err = waitContainerStop(ctx, c, time.Duration(timeout)*time.Second)
+		err = waitContainerStop(ctx, c, time.Duration(timeout)*time.Second, true)
 		if err == nil {
 			return nil
 		}
@@ -681,7 +685,7 @@ func (r *Runtime) StopContainer(ctx context.Context, c *Container, timeout int64
 		return fmt.Errorf("failed to stop container %s, %v", c.id, err)
 	}
 
-	return waitContainerStop(ctx, c, killContainerTimeout)
+	return waitContainerStop(ctx, c, killContainerTimeout, false)
 }
 
 // DeleteContainer deletes a container.
